Rename folder keyvalue-parser to kv-core
