+++ src/kv-core/kv-string-util.test.ts
