+++ src/kv-core/kv-string-util.ts
