+++ src/kv-core/kv-tokenizer.ts
